Aula 08/02

+++ biblioteca/src/main/java/br/com/fuctura/biblioteca/services/CategoriaService.java
@@ -1,8 +1,10 @@
 package br.com.fuctura.biblioteca.services;
 
+import br.com.fuctura.biblioteca.exceptions.ObjectNotFoundException;
 import br.com.fuctura.biblioteca.models.Categoria;
 import br.com.fuctura.biblioteca.repositories.CategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,26 +18,43 @@ public class CategoriaService {
 
     public Categoria findById(Integer id) {
         Optional<Categoria> cat = categoriaRepository.findById(id);
-        return cat.orElse(null);
+        if (cat.isPresent()) {
+            return cat.get();
+        }
+        throw new ObjectNotFoundException("Categoria não Encontrada");
+        //return cat.orElseThrow(() -> new ObjectNotFoundException("Categoria não Encontrada"));
     }
 
     public List<Categoria> findAll() {
         return categoriaRepository.findAll();
     }
 
-
     public Categoria save(Categoria categoria) {
+        findByGenero(categoria);
         Categoria cat = categoriaRepository.save(categoria);
         return cat;
     }
 
     public Categoria update(Categoria categoria) {
+        findById(categoria.getId());
+        findByGenero(categoria);
         Categoria cat = categoriaRepository.save(categoria);
         return cat;
     }
 
-    public void delete(Integer id){
-       categoriaRepository.deleteById(id);
+    public void delete(Integer id) {
+        Categoria cat = findById(id);
+        if (!cat.getLivros().isEmpty()) {
 
+            throw new DataIntegrityViolationException("Categoria possui livros, não pode ser deletada!");
+        }
+        categoriaRepository.deleteById(id);
+    }
+
+    private void findByGenero(Categoria categoria) {
+        Optional<Categoria> cat = categoriaRepository.findByGenero(categoria.getGenero());
+        if (cat.isPresent() && cat.get().getGenero().equals(categoria.getGenero())) {
+            throw new IllegalArgumentException("Já existe uma categoria com este gênero");
+        }
     }
 }
